experiments with analysing files for potential derived data types
outputs some likely candidates in a line-by-line format:

internal_varname({indices},...) where indices are a list of constants used in
that index position and/or * to indicate a non-constant.

a candidate array would likely have many constant indices and no * in a particular position.

a candidate array of arrays may have constant indices in one position while having a * in another position.